Move createTransformerWatcher
