+++ src/Project/functions/compileFiles.ts
@@ -3,10 +3,10 @@ import fs from "fs-extra";
 import { renderAST } from "LuauRenderer";
 import path from "path";
 import { checkRojoConfig } from "Project/functions/checkRojoConfig";
-import { createTransformerWatcher } from "Project/functions/createTransformerWatcher";
 import { transformPaths } from "Project/transformers/builtin/transformPaths";
 import { transformTypeReferenceDirectives } from "Project/transformers/builtin/transformTypeReferenceDirectives";
 import { createTransformerList, flattenIntoTransformers } from "Project/transformers/createTransformerList";
+import { createTransformerWatcher } from "Project/transformers/createTransformerWatcher";
 import { getPluginConfigs } from "Project/transformers/getPluginConfigs";
 import { getCustomPreEmitDiagnostics } from "Project/util/getCustomPreEmitDiagnostics";
 import { hasErrors } from "Project/util/hasErrors";
